+++ spec/whois/record/parser/responses/whois.nic.co/status_registered_spec.rb
@@ -21,6 +21,26 @@ describe Whois::Record::Parser::WhoisNicCo, "status_registered.expected" do
     described_class.new(part)
   end
 
+  describe "#domain" do
+    it do
+      subject.domain.should == "t.co"
+    end
+  end
+  describe "#domain_id" do
+    it do
+      subject.domain_id.should == "D740225-CO"
+    end
+  end
+  describe "#referral_whois" do
+    it do
+      lambda { subject.referral_whois }.should raise_error(Whois::PropertyNotSupported)
+    end
+  end
+  describe "#referral_url" do
+    it do
+      lambda { subject.referral_url }.should raise_error(Whois::PropertyNotSupported)
+    end
+  end
   describe "#status" do
     it do
       subject.status.should == ["serverDeleteProhibited", "serverTransferProhibited"]
@@ -52,6 +72,73 @@ describe Whois::Record::Parser::WhoisNicCo, "status_registered.expected" do
     it do
       subject.expires_on.should be_a(Time)
       subject.expires_on.should == Time.parse("2013-04-25 23:59:59 UTC")
+    end
+  end
+  describe "#registrar" do
+    it do
+      subject.registrar.should be_a(Whois::Record::Registrar)
+      subject.registrar.id.should           == "13"
+      subject.registrar.name.should         == "MELBOURNE IT LTD"
+    end
+  end
+  describe "#registrant_contacts" do
+    it do
+      subject.registrant_contacts.should be_a(Array)
+      subject.registrant_contacts.should have(1).items
+      subject.registrant_contacts[0].should be_a(Whois::Record::Contact)
+      subject.registrant_contacts[0].type.should          == Whois::Record::Contact::TYPE_REGISTRANT
+      subject.registrant_contacts[0].id.should            == "TWITTERREG2012"
+      subject.registrant_contacts[0].name.should          == "Twitter, Inc."
+      subject.registrant_contacts[0].organization.should  == "Twitter, Inc."
+      subject.registrant_contacts[0].address.should       == "1355 Market Street\nSuite 900"
+      subject.registrant_contacts[0].city.should          == "San Francisco"
+      subject.registrant_contacts[0].zip.should           == "94103"
+      subject.registrant_contacts[0].state.should         == "CA"
+      subject.registrant_contacts[0].country.should       == "United States"
+      subject.registrant_contacts[0].country_code.should  == "US"
+      subject.registrant_contacts[0].phone.should         == "+1.4152229670"
+      subject.registrant_contacts[0].fax.should           == "+1.4152220922"
+      subject.registrant_contacts[0].email.should         == "domains@twitter.com"
+    end
+  end
+  describe "#admin_contacts" do
+    it do
+      subject.admin_contacts.should be_a(Array)
+      subject.admin_contacts.should have(1).items
+      subject.admin_contacts[0].should be_a(Whois::Record::Contact)
+      subject.admin_contacts[0].type.should          == Whois::Record::Contact::TYPE_ADMIN
+      subject.admin_contacts[0].id.should            == "TWITTERADMIN2012"
+      subject.admin_contacts[0].name.should          == "Domain Admin"
+      subject.admin_contacts[0].organization.should  == "Twitter, Inc."
+      subject.admin_contacts[0].address.should       == "1355 Market Street\nSuite 900"
+      subject.admin_contacts[0].city.should          == "San Francisco"
+      subject.admin_contacts[0].zip.should           == "94103"
+      subject.admin_contacts[0].state.should         == "CA"
+      subject.admin_contacts[0].country.should       == "United States"
+      subject.admin_contacts[0].country_code.should  == "US"
+      subject.admin_contacts[0].phone.should         == "+1.4152229670"
+      subject.admin_contacts[0].fax.should           == "+1.4152220922"
+      subject.admin_contacts[0].email.should         == "domains@twitter.com"
+    end
+  end
+  describe "#technical_contacts" do
+    it do
+      subject.technical_contacts.should be_a(Array)
+      subject.technical_contacts.should have(1).items
+      subject.technical_contacts[0].should be_a(Whois::Record::Contact)
+      subject.technical_contacts[0].type.should          == Whois::Record::Contact::TYPE_TECHNICAL
+      subject.technical_contacts[0].id.should            == "TWITTERTECH2012"
+      subject.technical_contacts[0].name.should          == "Tech Admin"
+      subject.technical_contacts[0].organization.should  == "Twitter, Inc."
+      subject.technical_contacts[0].address.should       == "1355 Market Street\nSuite 900"
+      subject.technical_contacts[0].city.should          == "San Francisco"
+      subject.technical_contacts[0].zip.should           == "94103"
+      subject.technical_contacts[0].state.should         == "CA"
+      subject.technical_contacts[0].country.should       == "United States"
+      subject.technical_contacts[0].country_code.should  == "US"
+      subject.technical_contacts[0].phone.should         == "+1.4152229670"
+      subject.technical_contacts[0].fax.should           == "+1.4152220922"
+      subject.technical_contacts[0].email.should         == "domains-tech@twitter.com"
     end
   end
   describe "#nameservers" do